Split up code into core/ and ui/ modules
