+++ src/core/copy.rs
