+++ src/core/mod.rs
@@ -1,0 +1,7 @@
+pub mod initialize;
+mod address_finder;
+pub mod copy;
+#[cfg(target_os = "macos")]
+mod mac_maps;
+mod proc_maps;
+mod ruby_version;
